Stop relying on dead-reckoning mouse buttons for motion_notify_event.
Previously, for motion_notify_event, `event.attribute` was set by
checking the last button_press_event/button_release event.  This is
brittle for the same reason as to why we introduced `event.modifiers`
to improve over `event.key`, so introduce the more robust
`event.buttons` (see detailed discussion in the attribute docstring).

For a concrete example, consider e.g.

    from matplotlib import pyplot as plt
    from matplotlib.backends.qt_compat import QtWidgets

    def on_button_press(event):
        if event.button != 3:  # Right-click.
            return
        menu = QtWidgets.QMenu()
        menu.addAction("Some menu action", lambda: None)
        menu.exec(event.guiEvent.globalPosition().toPoint())

    fig = plt.figure()
    fig.canvas.mpl_connect("button_press_event", on_button_press)
    fig.add_subplot()
    plt.show()

(connecting a contextual menu on right button click) where a right click
while having selected zoom mode on the toolbar starts a zoom mode that
stays on even after the mouse release (because the mouse release event
is received by the menu widget, not by the main canvas).  This PR does
not fix the issue, but will allow a followup fix (where the
motion_notify_event associated with zoom mode will be able to first
check whether the button is indeed still pressed when the motion
occurs).

Limitations, on macOS only (everything works on Linux and Windows
AFAICT):
- tk only reports a single pressed button even if multiple buttons are
  pressed.
- gtk4 spams the terminal with Gtk-WARNINGs: "Broken accounting of
  active state for widget ..." on right-clicks only; similar
  issues appear to have been reported a while ago to Gtk
  (https://gitlab.gnome.org/GNOME/gtk/-/issues/3356 and linked issues)
  but it's unclear whether any action was taken on their side.

(Alternatively, some GUI toolkits have a "permissive" notion of
drag events defined as mouse moves with a button pressed, which we could
use as well to define event.button{,s} for motion_notify_event, but
e.g. Qt attaches quite heavy semantics to drags which we probably don't
want to bother with.)